fe_batch_invert: Use size_t for n to avoid negative n
Cypher Stack 4.3.2
https://github.com/cypherstack/fcmp-impl-audit/blob/d7de33d08c48a3c249629e47f71ea72c938a3404/FCMP_Code_Implementation_Audit.pdf

+++ src/crypto/crypto-ops.h
@@ -30,6 +30,7 @@
 
 #pragma once
 
+#include <stddef.h>
 #include <stdint.h>
 
 /* From fe.h */
@@ -174,7 +175,7 @@ If `out` and `in` overlap, it will cause undefined output.
 
 No 0 fe's are expected for `in`, otherwise fails.
 **/
-int fe_batch_invert(fe *out, const fe *in, const int n);
+int fe_batch_invert(fe *out, const fe *in, const size_t n);
 void fe_mul(fe out, const fe, const fe);
 void fe_0(fe h);
